refactor: moved Enums and EnumsOptions for cleaner code
